requirements.txt

rasterio>=1.3
numpy>=1.24
Pillow>=10.0

